更新内容： (#310)
* 修复一些已知问题

* 1.优化处理POST 数据时的性能
2.修复修改密码时能设置非法字符问题
3.删除残余代码
4.系统添加限制PHP7.0以下版本不能使用

* 更新内容：
1.文件管理支持打印与导出更多数据
2.管理页面支持显示php扩展信息

* 更新内容：
大幅提升数据表格加载体验
数据表格添加更多可选分页行数

* 支持更多主题色设置

* 修复pdf阅读器应用能打开无预览权限的文件问题

+++ dzz/share/ajax.php
@@ -9,27 +9,37 @@
 if (!defined('IN_DZZ')) {
 	exit('Access Denied');
 }
+if(!$_G['uid']) exit(json_encode(array('error' => lang('no_login_operation'))));
 if ($_GET['do'] == 'delete') {
 	$sids = $_GET['sids'];
     $return = array();
     foreach($sids as $v){
+		if(!$_G['adminid']) {
+			$suid = DB::result_first("select uid from %t where id=%s", array('shares', $v));
+			if($suid !== $_G['uid']) {
+				exit(json_encode(array('error' => '该分享文件不存在或您没有权限')));
+			}
+		}
         $result = C::t('shares')->delete_by_id($v);
-        if($result['success']){
-            $return['msg'][$v]=$result;
-        }elseif ($result['error']){
-            $return['msg'][$v] = $result['error'];
+        if ($result['error']){
+			exit(json_encode(array('error' => $result['error'])));
         }
     }
-    exit(json_encode($return));
+	exit(json_encode(array('success' => true)));
 } elseif ($_GET['do'] == 'forbidden' && $_G['adminid']) {
 	$sids = $_GET['sids'];
+	if (!$sids) {
+		exit(json_encode(array('error' => '非法操作')));
+	}
 	if ($_GET['flag'] == 'forbidden') {
 		$status = -4;
+		$msg = lang('share_screen_success');
 	} else {
 		$status = 0;
+		$msg = lang('cancel_shielding_success');
 	}
-	if ($sids && C::t('shares') -> update($sids, array('status' => $status))) {
-		exit(json_encode(array('msg' => 'success')));
+	if ($sids && C::t('shares')->update($sids, array('status' => $status))) {
+		exit(json_encode(array('msg' => $msg)));
 	} else {
 		exit(json_encode(array('error' => lang('share_screen_failure'))));
 	}
